Use a flag to enable or disable the sidebar
We default to showing the sidebar, but allow overriding that default to
remove it. We use this on the guides index page.

To save complexity of laying out the contribute button, we also hide
that if we hide the sidebar. We may want to revisit that decision if we
end up using it in other places.